+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+### Fixed
+
+-   [Patch] Prepend `$` to duration and easing SCSS tokens.
+-   [Patch] Remove `ms` from duration JS tokens.
+
 ## 10.1.3 - 2019-12-10
 
 ### Fixed